Use a GenServer instead of Agents for isolated test process uploads (#4)
## Problem 

Currently we need to make tests non-async and start instances of the test adapter per test to keep our test uploads isolated. This is fine for small parts of the codebase but scaling to multiple schemas exacerbates these issues.

## Solution

Use a GenServer which does book-keeping of processes performing Uploads.

This allows us to remove calls to `start_supervised` and use one GenServer instance for testing, which allows us to remain async safe, as only allowed processes will be able to access their own uploads. Eventually we may want to introduce a global mode like Mox, but for now this should be sufficient.