Casting an Upload of the same filename should generate a change (#5)
## Summary

Currently if an upload is attempted to be cast twice and it has the same filename, it will not be uploaded due to Ecto thinking there is no real change. From Ecto's perspective the column data is the same (The URL) so why should it make a change? This is an edge-case that was not tested for when we adjusted the library to not generate a unique filename on image creation.

## Solution

Always put the change regardless of the filename being the same.